server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Daniel Le <ledanieln@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "1.0.2"
Flask-Cors = "3.0.7"
google-cloud-bigquery = "^1.24.0"
python-dotenv = "^0.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
